requirements.txt
langchain==0.1.16
langchain-openai==0.1.3
langchain-core==0.1.45
langchain-community==0.0.34
langchain-groq==0.1.3
langchain-objectbox==0.1.0a0
langchainhub==0.1.15
langserve==0.1.0
python-dotenv==1.0.1
streamlit==1.33.0
fastapi==0.110.1
uvicorn==0.29.0
sse-starlette==2.0.0
bs4==0.0.2
beautifulsoup4==4.12.3
pypdf==4.2.0
PyPDF2==3.0.1
chromadb==0.4.24
faiss-cpu==1.8.0
groq==0.5.0
cassio==0.1.6
wikipedia==1.4.0
arxiv==2.1.0
sentence-transformers==2.7.0